tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.roundrect.radius.toomany.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.roundrect.radius.toomany</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.roundrect.radius.toomany</h1>
<p class="desc">Check that roundRect throws an IndeSizeError if radii has more than four items.</p>


<script>
var t = async_test("Check that roundRect throws an IndeSizeError if radii has more than four items.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_throws_js(RangeError, () => { ctx.roundRect(0, 0, 100, 50, [0, 0, 0, 0, 0])});
  t.done();

});
</script>
